Improves UI with Tailwind Merge update & styling
Updates tailwind-merge dependency to the latest version.

Refactors the UI with new styles for hero patterns and gradient text to improve visual appeal and user experience.
Moves the Navbar component to ResizableNavbar for improved responsiveness on different screen sizes, also placing it correctly so the content does not overlap the navbar.